fixed double folder bugs

+++ src/eg/edu/alexu/csd/oop/cs71/jdbc/src/Statement.java
@@ -18,6 +18,7 @@ public class Statement implements java.sql.Statement {
     ArrayList<String> Queries = new ArrayList<>();
     Properties info;
     boolean closeState = false;
+    DBLogger dbLogger =DBLogger.getInstance();
 
     public Statement(Connection connection, Properties info) {
         this.connection = connection;
@@ -37,8 +38,10 @@ public class Statement implements java.sql.Statement {
             ArrayList<String> types = facade.getColumnTypes();
             String tableName = fm.getTableName(sql);
             Resultset rs = new Resultset(tableName, data, types);
+            dbLogger.addLog("fine","Select Query executed");
             return rs;
         }
+        dbLogger.addLog("severe","Select Query failed");
         throw new SQLException("Invalid Query");
     }
 
@@ -49,8 +52,10 @@ public class Statement implements java.sql.Statement {
         }
         ValidationInterface SQLvalidation = new SQLBasicValidation();
         if (SQLvalidation.validateQuery(sql)) {
+            dbLogger.addLog("fine","Update Query executed");
             return (int) facade.parse(sql);
         }
+        dbLogger.addLog("Severe","Update Query Failed!");
         throw new SQLException("Invalid Query");
     }
 
@@ -58,6 +63,7 @@ public class Statement implements java.sql.Statement {
     public void close() throws SQLException {
         connection = null;
         closeState = true;
+        dbLogger.addLog("fine","Statement Closed");
     }
 
     @Override
@@ -146,17 +152,23 @@ public class Statement implements java.sql.Statement {
                         String tempo=temp[1].substring(1,temp[1].length()-1);
                         temp[1]=tempo;
                     }
+                    if(!temp[1].contains("\\")){
+                        facade.parse(sql);
+                        return true;
+                    }
                     String sql2;
                     if (checker.contains("CREATE"))
                         sql2 = "create database " + temp[1] + "\\" + command[2];
                     else
                         sql2 = "drop database " + temp[1] + "\\" + command[2];
                     facade.parse(sql2);
+                    dbLogger.addLog("fine","Create|Drop Query executed");
                 } else return (boolean) facade.parse(sql);
                 return true;
             }
             return false;
         } else {
+            dbLogger.addLog("Severe","Invalid Query");
             throw new SQLException("Invalid Query");
         }
     }
@@ -211,9 +223,15 @@ public class Statement implements java.sql.Statement {
         if (isClosed()) {
             throw new SQLException();
         }
-        if(sql.contains("update")||sql.contains("insert")||sql.contains("delete"))
-        Queries.add(sql);
-        else throw new SQLException("can't add a non-update query");
+
+        if(sql.contains("update")||sql.contains("insert")||sql.contains("delete")){
+            Queries.add(sql);
+            dbLogger.addLog("fine","Query Added to Batch");
+        }
+        else{
+            dbLogger.addLog("Severe","Failed to add to batch");
+            throw new SQLException("can't add a non-update query");
+        }
     }
 
     @Override
@@ -222,12 +240,16 @@ public class Statement implements java.sql.Statement {
             throw new SQLException();
         }
         Queries.clear();
+        dbLogger.addLog("fine","Batch cleared");
     }
 
     @Override
     public int[] executeBatch() throws SQLException {
         ArrayList<Integer> rowsAffected =new ArrayList<Integer>();
-        if(Queries.size()==0)throw new SQLException("Batch is empty");
+        if(Queries.size()==0){
+            dbLogger.addLog("Severe","Failed to execute batch");
+            throw new SQLException("Batch is empty");
+        }
         for(int i=0;i<Queries.size();i++)
         {
             if(Queries.get(i).contains("update")||Queries.get(i).contains("insert")||Queries.get(i).contains("delete"))
@@ -235,7 +257,9 @@ public class Statement implements java.sql.Statement {
         }
         int[] rowsAffected1=new int[rowsAffected.size()];
         for(int i=0;i<rowsAffected.size();i++)rowsAffected1[i]=rowsAffected.get(i);
+        dbLogger.addLog("fine","Batch Executed");
         return rowsAffected1;
+
     }
 
     @Override
